refactor!: tenant_id 类型统一 string——消除哨兵值 0 的语义重载
## 根因

viewer.Context.TenantID() 为 uint64，而 bald 生态（contextx/authn/jwt/audit）
全为 string。crudbridge 桥接时 strconv.ParseUint 转换，非数字租户 ID
（如 bald-admin 的 "t-default"）解析失败被**静默留 0**——而
IsPlatformContext() 定义为 TenantIDValue == 0，于是：

    "t-default" → 解析失败 → 0 → 平台视图 → EnforceTenant 放行 → 租户隔离静默失效

**uint64 的 0 同时表达「平台视图」与「解析失败」两种互斥含义**（哨兵值语义
重载）。这是安全缺陷：该隔离的没隔离，且不可见。

## 修复

统一为 string——空串即显式的「无租户上下文」，无解析失败路径。

- viewer: Context.TenantID() uint64→string；EnforcementDecision.TenantID 同步；
  ScopedModel Get/SetTenantID 同步
- 各 ORM mixin（gorm/clickhouse/doris/mongodb）：TenantID *uint32→*string；
  gorm 列类型 int unsigned→varchar(64)
- entgo: 去掉伪泛型 （IDT 从未用于字段类型，
  硬编码 field.Uint32）；字段改 field.String；go generate 重生成
- audit: Entry.TenantID uint64→string
- entgo/interceptor + rule: 去泛型参数

## 验证

- 6 个 module（viewer/audit/gorm/clickhouse/doris/mongodb）编译通过
- entgo 核心包（mixin/rule/interceptor/ent）编译通过并重生成
- 生成代码已确认：TenantID *string / predicate.User 收 string

## 已知（预先存在，非本次引入）

entgo/{field,sorting,filter} 三包因 ent.CheckColumn 未定义而编译失败——
该符号在 26d1829（仓库创建提交）即存在，entgo.io/ent@v0.14.6 无此函数。

破坏性变更：viewer.Context 是公开契约，实现者需同步签名。

+++ doris/tenant_enforce.go
@@ -95,7 +95,7 @@ func enforceTenantOnBatchItem[ENTITY any](ctx context.Context, rv reflect.Value,
 			return err
 		}
 		if dec.Enforce {
-			m["tenant_id"] = uint32(dec.TenantID)
+			m["tenant_id"] = dec.TenantID
 		}
 		return nil
 	case reflect.Struct:
